chore: webrtc tests are multi platform
